requirements.txt
langchain==0.1.16
langchain_google_genai==1.0.2
protobuf==5.26.1
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.31.1
google-generativeai
faiss-cpu